+++ src/iexec-app.js
@@ -7,6 +7,7 @@ const {
   addWalletLoadOptions,
   computeWalletLoadOptions,
   computeTxOptions,
+  checkUpdate,
   handleError,
   desc,
   option,
@@ -35,6 +36,7 @@ const init = cli.command('init');
 addGlobalOptions(init);
 addWalletLoadOptions(init);
 init.description(desc.initObj(objName)).action(async (cmd) => {
+  await checkUpdate(cmd);
   const spinner = Spinner(cmd);
   try {
     const walletOptions = await computeWalletLoadOptions(cmd);
@@ -64,6 +66,7 @@ deploy
   .option(...option.txGasPrice())
   .description(desc.deployObj(objName))
   .action(async (cmd) => {
+    await checkUpdate(cmd);
     const spinner = Spinner(cmd);
     try {
       const walletOptions = await computeWalletLoadOptions(cmd);
@@ -113,6 +116,7 @@ show
   .option(...option.user())
   .description(desc.showObj(objName))
   .action(async (cliAddressOrIndex, cmd) => {
+    await checkUpdate(cmd);
     const spinner = Spinner(cmd);
     try {
       const walletOptions = await computeWalletLoadOptions(cmd);
@@ -155,6 +159,7 @@ count
   .option(...option.user())
   .description(desc.countObj(objName))
   .action(async (cmd) => {
+    await checkUpdate(cmd);
     const spinner = Spinner(cmd);
     const walletOptions = await computeWalletLoadOptions(cmd);
     const keystore = Keystore(