+++ src/iexec-deal.js
@@ -6,6 +6,7 @@ const {
   addGlobalOptions,
   addWalletLoadOptions,
   computeWalletLoadOptions,
+  checkUpdate,
   handleError,
   desc,
   option,
@@ -28,6 +29,7 @@ show
   .option(...option.chain())
   .description(desc.showObj(objName))
   .action(async (dealid, cmd) => {
+    await checkUpdate(cmd);
     const spinner = Spinner(cmd);
     try {
       const chain = await loadChain(cmd.chain, Keystore({ isSigner: false }), {