+++ src/iexec-order.js
@@ -8,6 +8,7 @@ const {
   addWalletLoadOptions,
   computeWalletLoadOptions,
   computeTxOptions,
+  checkUpdate,
   handleError,
   desc,
   option,
@@ -45,6 +46,7 @@ init
   .option(...option.chain())
   .description(desc.initObj(objName))
   .action(async (cmd) => {
+    await checkUpdate(cmd);
     const spinner = Spinner(cmd);
     try {
       const initAll = !(
@@ -105,6 +107,7 @@ sign
   .option(...option.chain())
   .description(desc.sign())
   .action(async (cmd) => {
+    await checkUpdate(cmd);
     const spinner = Spinner(cmd);
     try {
       const signAll = !(
@@ -274,6 +277,7 @@ fill
   .option(...option.fillRequestParams())
   .description(desc.fill(objName))
   .action(async (cmd) => {
+    await checkUpdate(cmd);
     const spinner = Spinner(cmd);
     try {
       const walletOptions = await computeWalletLoadOptions(cmd);
@@ -536,6 +540,7 @@ publish
   .option(...option.chain())
   .description(desc.publish(objName))
   .action(async (cmd) => {
+    await checkUpdate(cmd);
     const spinner = Spinner(cmd);
     try {
       if (!(cmd.app || cmd.dataset || cmd.workerpool || cmd.request)) {
@@ -598,6 +603,7 @@ unpublish
   .option(...option.force())
   .description(desc.unpublish(objName))
   .action(async (cmd) => {
+    await checkUpdate(cmd);
     const spinner = Spinner(cmd);
     try {
       if (!(cmd.app || cmd.dataset || cmd.workerpool || cmd.request)) {
@@ -680,6 +686,7 @@ cancel
   .option(...option.force())
   .description(desc.cancel(objName))
   .action(async (cmd) => {
+    await checkUpdate(cmd);
     const spinner = Spinner(cmd);
     try {
       if (!(cmd.app || cmd.dataset || cmd.workerpool || cmd.request)) {
@@ -732,6 +739,7 @@ show
   .option(...option.chain())
   .description(desc.showObj(objName, 'marketplace'))
   .action(async (cmd) => {
+    await checkUpdate(cmd);
     const spinner = Spinner(cmd);
     try {
       if (!(cmd.app || cmd.dataset || cmd.workerpool || cmd.request)) {