+++ src/iexec-wallet.js
@@ -16,6 +16,7 @@ const {
   addWalletLoadOptions,
   computeWalletLoadOptions,
   computeTxOptions,
+  checkUpdate,
   handleError,
   help,
   Spinner,
@@ -36,6 +37,7 @@ create
   .option(...option.forceCreate())
   .description(desc.createWallet())
   .action(async (cmd) => {
+    await checkUpdate(cmd);
     const spinner = Spinner(cmd);
     try {
       const force = cmd.force || false;
@@ -62,6 +64,7 @@ importPk
   .option(...option.forceCreate())
   .description(desc.importWallet())
   .action(async (privateKey, cmd) => {
+    await checkUpdate(cmd);
     const spinner = Spinner(cmd);
     try {
       const force = cmd.force || false;
@@ -90,6 +93,7 @@ show
   .option(...option.showPrivateKey())
   .description(desc.showObj(objName, 'address'))
   .action(async (address, cmd) => {
+    await checkUpdate(cmd);
     const spinner = Spinner(cmd);
     try {
       const walletOptions = await computeWalletLoadOptions(cmd);
@@ -149,6 +153,7 @@ getEth
   .option(...option.chain())
   .description(desc.getETH())
   .action(async (cmd) => {
+    await checkUpdate(cmd);
     const spinner = Spinner(cmd);
     try {
       const walletOptions = await computeWalletLoadOptions(cmd);
@@ -185,6 +190,7 @@ getRlc
   .option(...option.chain())
   .description(desc.getRLC())
   .action(async (cmd) => {
+    await checkUpdate(cmd);
     const spinner = Spinner(cmd);
     try {
       const walletOptions = await computeWalletLoadOptions(cmd);
@@ -224,6 +230,7 @@ sendETH
   .option(...option.force())
   .description(desc.sendETH())
   .action(async (amount, cmd) => {
+    await checkUpdate(cmd);
     const spinner = Spinner(cmd);
     try {
       const walletOptions = await computeWalletLoadOptions(cmd);
@@ -261,6 +268,7 @@ sendRLC
   .option(...option.force())
   .description(desc.sendRLC())
   .action(async (amount, cmd) => {
+    await checkUpdate(cmd);
     const spinner = Spinner(cmd);
     try {
       const walletOptions = await computeWalletLoadOptions(cmd);
@@ -300,6 +308,7 @@ sweep
   .option(...option.force())
   .description(desc.sweep())
   .action(async (cmd) => {
+    await checkUpdate(cmd);
     const spinner = Spinner(cmd);
     try {
       const walletOptions = await computeWalletLoadOptions(cmd);